# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EmployeeStatus = typeof EmployeeStatus[keyof typeof EmployeeStatus];


export const EmployeeStatus = {
  active: 'active',
  resigned: 'resigned',
} as const;

export interface Employee {
  id: number;
  name: string;
  department: string;
  role: string;
  joinedAt: string;
  annualAllowance: number;
  usedDays: number;
  pendingDays: number;
  remainingDays: number;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  employeeNumber?: string | null;
  position: string;
  status: EmployeeStatus;
  /** @nullable */
  resignedAt?: string | null;
  ordinaryHourlyWage: number;
}

export interface EmployeeRegistrationInput {
  /** @minLength 1 */
  companyName: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
}

export type AdminEmployeeInputRole = typeof AdminEmployeeInputRole[keyof typeof AdminEmployeeInputRole];


export const AdminEmployeeInputRole = {
  직원: '직원',
  관리자: '관리자',
} as const;

export interface AdminEmployeeInput {
  /** @minLength 1 */
  employeeNumber: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  department: string;
  position: string;
  role: AdminEmployeeInputRole;
  joinedAt: string;
  /** @minimum 0 */
  ordinaryHourlyWage: number;
}

export type AdminEmployeeUpdateInputRole = typeof AdminEmployeeUpdateInputRole[keyof typeof AdminEmployeeUpdateInputRole];


export const AdminEmployeeUpdateInputRole = {
  직원: '직원',
  관리자: '관리자',
} as const;

export interface AdminEmployeeUpdateInput {
  /** @minLength 1 */
  employeeNumber?: string;
  /** @minLength 1 */
  name?: string;
  /** @minLength 3 */
  email?: string;
  /** @minLength 1 */
  department?: string;
  position?: string;
  role?: AdminEmployeeUpdateInputRole;
  joinedAt?: string;
  /** @minimum 0 */
  ordinaryHourlyWage?: number;
}

export interface ResignEmployeeInput {
  resignedAt: string;
}

export interface LeaveBalance {
  employeeId: number;
  annualAllowance: number;
  usedDays: number;
  pendingDays: number;
  remainingDays: number;
}

export type LeaveRequestStatus = typeof LeaveRequestStatus[keyof typeof LeaveRequestStatus];


export const LeaveRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export type LeaveType = typeof LeaveType[keyof typeof LeaveType];


export const LeaveType = {
  annual: 'annual',
  half_day: 'half_day',
  quarter_day: 'quarter_day',
  early_leave: 'early_leave',
  outing: 'outing',
  wedding_funeral: 'wedding_funeral',
  paid_leave: 'paid_leave',
  sick_leave: 'sick_leave',
  substitute: 'substitute',
  absence: 'absence',
  sabbatical: 'sabbatical',
} as const;

export type TimeSlot = typeof TimeSlot[keyof typeof TimeSlot];


export const TimeSlot = {
  start: 'start',
  end: 'end',
} as const;

/**
 * @nullable
 */
export type LeaveRequestTimeSlot = typeof LeaveRequestTimeSlot[keyof typeof LeaveRequestTimeSlot] | null;


export const LeaveRequestTimeSlot = {
  start: 'start',
  end: 'end',
} as const;

export interface LeaveRequest {
  id: number;
  employeeId: number;
  employeeName: string;
  department: string;
  leaveType: LeaveType;
  /** @nullable */
  timeSlot?: LeaveRequestTimeSlot;
  startDate: string;
  endDate: string;
  days: number;
  reason: string;
  status: LeaveRequestStatus;
  createdAt: string;
  /** @nullable */
  processedAt?: string | null;
  /** @nullable */
  rejectionReason?: string | null;
}

/**
 * @nullable
 */
export type LeaveRequestInputTimeSlot = typeof LeaveRequestInputTimeSlot[keyof typeof LeaveRequestInputTimeSlot] | null;


export const LeaveRequestInputTimeSlot = {
  start: 'start',
  end: 'end',
} as const;

export interface LeaveRequestInput {
  employeeId: number;
  leaveType: LeaveType;
  /** @nullable */
  timeSlot?: LeaveRequestInputTimeSlot;
  startDate: string;
  endDate: string;
  /** @minimum 0.25 */
  days: number;
  /** @minLength 1 */
  reason: string;
}

export interface AbsenceInput {
  employeeId: number;
  startDate: string;
  endDate: string;
  /** @minimum 0.5 */
  days: number;
  /** @minLength 1 */
  reason: string;
}

export interface LeaveRequestStatusInput {
  status: LeaveRequestStatus;
  /** @nullable */
  rejectionReason?: string | null;
}

export interface UpcomingLeave {
  id: number;
  employeeName: string;
  leaveType: LeaveType;
  startDate: string;
  endDate: string;
  days: number;
}

export interface MonthlyLeaveUsage {
  month: string;
  days: number;
}

export interface DashboardSummary {
  year: number;
  annualAllowance: number;
  usedDays: number;
  remainingDays: number;
  pendingDays: number;
  employeeCount: number;
  pendingRequestCount: number;
  approvedRequestCount: number;
  upcomingLeaves: UpcomingLeave[];
  monthlyUsage: MonthlyLeaveUsage[];
}

export type SearchParameter = string;

export type StatusParameter = LeaveRequestStatus;

export type YearParameter = number;

export type MonthParameter = number;

export type EmployeeIdQueryParameter = number;

export type GetEmployeesParams = {
search?: SearchParameter;
};

export type GetAdminEmployeesParams = {
search?: SearchParameter;
};

export type GetLeaveRequestsParams = {
status?: StatusParameter;
search?: SearchParameter;
year?: YearParameter;
/**
 * @minimum 1
 * @maximum 12
 */
month?: MonthParameter;
employeeId?: EmployeeIdQueryParameter;
};