# lib/api-zod/src/generated/types/adminEmployeeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminEmployeeInputRole } from './adminEmployeeInputRole';

export interface AdminEmployeeInput {
  /** @minLength 1 */
  employeeNumber: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  department: string;
  position: string;
  role: AdminEmployeeInputRole;
  joinedAt: Date;
  /** @minimum 0 */
  ordinaryHourlyWage: number;
}