# lib/api-zod/src/generated/types/adminEmployeeUpdateInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminEmployeeUpdateInputRole } from './adminEmployeeUpdateInputRole';

export interface AdminEmployeeUpdateInput {
  /** @minLength 1 */
  employeeNumber?: string;
  /** @minLength 1 */
  name?: string;
  /** @minLength 3 */
  email?: string;
  /** @minLength 1 */
  department?: string;
  position?: string;
  role?: AdminEmployeeUpdateInputRole;
  joinedAt?: Date;
  /** @minimum 0 */
  ordinaryHourlyWage?: number;
}