# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStatus } from './employeeStatus';

export interface Employee {
  id: number;
  name: string;
  department: string;
  role: string;
  joinedAt: Date;
  annualAllowance: number;
  usedDays: number;
  pendingDays: number;
  remainingDays: number;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  employeeNumber?: string | null;
  position: string;
  status: EmployeeStatus;
  /** @nullable */
  resignedAt?: Date | null;
  ordinaryHourlyWage: number;
}